move countChunks and forAllChunks to iteration.h
